fixed fog first terrain approach

+++ src/TerrainShader.h
@@ -18,9 +18,9 @@ private:
 	RGBImage* NoiseMapImage;
 	void initNoiseTexture();
 
-    int WorldWidth = GameSettings::WORLD_WIDTH;
-    int WorldDepth = GameSettings::WORLD_DEPTH;
-    int MountainWidth = GameSettings::MOUNTAIN_WIDTH;
+    float WorldWidth = GameSettings::WORLD_WIDTH;
+    float WorldDepth = GameSettings::WORLD_DEPTH;
+    float MountainWidth = GameSettings::MOUNTAIN_WIDTH;
 
 	Texture* NoiseMapTexture = Texture::defaultTex();
